run pre-commit hooks to organize code

+++ orders/models.py
@@ -1,4 +1,5 @@
 from django.db import models
+
 from accounts.models import CustomUser
 from products.models import Product, TimeStampedModel
 
@@ -8,19 +9,21 @@ class Payment(TimeStampedModel):
     payment_id = models.CharField(max_length=100)
     payment_method = models.CharField(max_length=100)
     amount_paid = models.CharField(max_length=100)
-    status = models.CharField(max_length=100)    
+    status = models.CharField(max_length=100)
 
 
 class Order(TimeStampedModel):
     STATUS = (
-        ('New', 'New'),
-        ('Accepted', 'Accepted'),
-        ('Completed', 'Completed'),
-        ('Cancelled', 'Cancelled'),
+        ("New", "New"),
+        ("Accepted", "Accepted"),
+        ("Completed", "Completed"),
+        ("Cancelled", "Cancelled"),
     )
 
     user = models.ForeignKey(CustomUser, on_delete=models.SET_NULL, null=True)
-    payment = models.ForeignKey(Payment, on_delete=models.SET_NULL, blank=True, null=True)
+    payment = models.ForeignKey(
+        Payment, on_delete=models.SET_NULL, blank=True, null=True
+    )
     order_number = models.CharField(max_length=20)
     mobile = models.CharField(max_length=15)
     email = models.EmailField(max_length=50)
@@ -31,20 +34,22 @@ class Order(TimeStampedModel):
     city = models.CharField(max_length=50)
     order_note = models.CharField(max_length=100, blank=True)
     order_total = models.FloatField()
-    status = models.CharField(max_length=10, choices=STATUS, default='New')
+    status = models.CharField(max_length=10, choices=STATUS, default="New")
     is_ordered = models.BooleanField(default=False)
 
     def full_address(self):
-        return f'{self.address_line_1} {self.address_line_2}'
+        return f"{self.address_line_1} {self.address_line_2}"
 
 
 class OrderProduct(TimeStampedModel):
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
-    payment = models.ForeignKey(Payment, on_delete=models.SET_NULL, blank=True, null=True)
+    payment = models.ForeignKey(
+        Payment, on_delete=models.SET_NULL, blank=True, null=True
+    )
     user = models.ForeignKey(CustomUser, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
     quantity = models.IntegerField()
-    product_price = models.FloatField()
+    product_price = models.DecimalField()
     ordered = models.BooleanField(default=False)
 
     def __str__(self):
